+++ Changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- fixed error when trying to overwrite existing specs
+
 ## 1.3.0 - 2022-08-30
 
 ### Added